requirements.txt
python-dotenv
requests
spotipy
flask
os